refactor: extract bridge.js from main.js (BAT-195)
Move androidBridgeCall() (~45 lines) into dedicated bridge.js module.
Android Bridge HTTP client for local calls on port 8765.
Depends on config.js for BRIDGE_TOKEN and log. No logic changes.
